reik_finance_site/requirements.txt
Django==5.0.6
djangorestframework==3.15
django-cors-headers==3.14.0
requests==2.32.3
boto3==1.35.0
plaid-python==31.0.0
python-dotenv